pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zenmarket-ai"
version = "1.0.0"
description = "AI-powered financial intelligence and market analysis system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TechNatool", email = "contact@technatool.com"}
]
keywords = ["finance", "ai", "trading", "market-analysis", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pytz>=2024.1",
    "yfinance>=0.2.36",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "feedparser>=6.0.11",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "markdown2>=2.4.12",
    "weasyprint>=61.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
zenmarket = "src.main:main"

[project.urls]
Homepage = "https://github.com/TechNatool/zenmarket-ai"
Repository = "https://github.com/TechNatool/zenmarket-ai"
Issues = "https://github.com/TechNatool/zenmarket-ai/issues"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv", "venv"]
